+++ components/Content/Planets/Sectors/SectorSetup.tsx
