+++ components/Content/Planets/Sectors/StructureCreate.tsx
